+++ lib/smpp/pdu/submit_base.rb
@@ -1,0 +1,8 @@
+# encoding: UTF-8
+
+# Base class for the SubmitSm and the SubmitMulti classes
+# Author: Jorge Piera Llodra(jorge[at]pierallodra.es)
+class Smpp::Pdu::SubmitBase < Smpp::Pdu::Base
+  
+  
+end
